Opdatering af koordinatesystemerne sker ift. dent nye interface. tilføjet visning af interpolations-spline (viser sig at være zig-zag agtigt så den skal splittes op i 2)fjernet Edit fra hovedmenuen. Den var tom. tilføje View/Show interpolation til menuen Tilføjet editbox max SearchE-iterations så den kan sættes dynamisk tilføjet visning af sub-state-felter (decrM,hi/lo precision/interpolation done) samt main-iteration Splittet BigRealRemes2.cpp op i flere filer. Den var blevet lidt stor Bruger rProd(...,..,m_digits) i til af lave A-mtricen i findCoefficients istedet for * Omdøbt variablene m_left,m_right til m_domain som er et bigRealInterval (NumberInterval<BigReal>)

+++ Math/IRemes/MultiExtremaFinder.cpp
@@ -61,7 +61,7 @@ void MultiExtremaFinder::waitUntilAllThreadsTerminated() {
   int runningCount;
   for(;;) {
     runningCount = 0;
-    for(int i = 0; i < m_threads.size(); i++) {
+    for(size_t i = 0; i < m_threads.size(); i++) {
       if(m_threads[i]->stillActive()) {
         runningCount++;
       }
@@ -85,7 +85,7 @@ void MultiExtremaFinder::execute() {
     thr->start();
   }
   try {
-    for(int count = 0; count < m_allJobs.size(); count++) {
+    for(size_t count = 0; count < m_allJobs.size(); count++) {
       ExtremumResult result(&DEFAULT_DIGITPOOL);
       result = resultQueue.get();
       m_remes.setExtremum(result.m_index, result);
